Made implementation match the readme
