+++ twyla/kubedeploy/__init__.py
@@ -1,37 +1,14 @@
-import json
 import os
 import shutil
 import sys
 import tempfile
 
 import click
-import colorama
-import kubernetes.client
-import pip
-import yaml
-
 import git
-
-from .kube import Kube, KubeException
-from . import docker_helpers
-
-
-# Use a colorized prompt to differentiate output of this script from output
-# that is generated by called programs and libraries
-colorama.init(autoreset=True)
-PROMPT = '>> '
-
-
-def prompt(msg: str, indent: int=0):
-    indentation = ' ' * indent
-    sys.stdout.write(colorama.Fore.GREEN + PROMPT + indentation)
-    print(msg)
-
-
-def error_prompt(msg: str):
-    sys.stdout.write(colorama.Fore.RED + PROMPT)
-    print(msg)
-
+import pip
+from twyla.kubedeploy import docker_helpers
+from twyla.kubedeploy.kube import Kube
+from twyla.kubedeploy.prompt import error_prompt, prompt
 
 
 def download_requirements(force: bool=False):
@@ -57,8 +34,6 @@ def download_requirements(force: bool=False):
         deps = [line for line in f if line.startswith('git+ssh')]
     pip.main(['download', '-q', '--dest', tmp, *deps])
     shutil.move(tmp, dest)
-
-
 
 
 def head_of(working_directory: str, branch: str=None, local: bool=False) -> str:
@@ -125,19 +100,6 @@ def head_of(working_directory: str, branch: str=None, local: bool=False) -> str:
     return repo.git.rev_parse(remote_refs[0].commit, short=8)
 
 
-def load_options(base_path):
-    """Load the options for a service deployment. base_path is the
-    directory in which the command is called."""
-    default_service_name = base_path.rstrip(os.sep).rsplit(os.sep, 1)[-1]
-    default_options = {'namespace': 'default',
-                       'service_name': default_service_name}
-    config_path = os.path.join(base_path, '.kubedeploy')
-    if os.path.isfile(config_path):
-        with open(config_path, 'r') as config_file:
-            default_options.update(yaml.load(config_file.read()))
-    return default_options
-
-
 @click.group()
 @click.pass_context
 def cli(ctx: click.Context):
@@ -145,8 +107,13 @@ def cli(ctx: click.Context):
 
 
 @cli.command()
-@click.option('--registry')
-@click.option('--image')
+@click.option('--registry', help='Docker registry name.', required=True)
+@click.option('--image', help='Docker image name.',
+              required=True)
+@click.option('--name', help='Name of the deployment.',
+              required=True)
+@click.option('--namespace', help='The destination namespace in the cluster.',
+              default='default')
 @click.option('--branch', help='The git branch to deploy. Defaults to master.',
               default='master')
 @click.option('--version', help='Version of API to build and deploy. Will'
@@ -158,25 +125,19 @@ def cli(ctx: click.Context):
               ' service will be used to create, push, and deploy a Docker'
               ' image.',
               default=False)
-def deploy(registry: str, image: str, branch: str, version: str,
-           local: bool, dry: bool):
+def deploy(registry: str, image: str, name: str, namespace: str, branch: str,
+           version: str, local: bool, dry: bool):
     working_directory = os.getcwd()
-    options = load_options(working_directory)
-    registry = registry or options['registry']
-    image = image or options['service_name']
     if local:
         # Reset branch when using local.
         branch = None
     if version is None:
         version = head_of(working_directory, branch, local=local)
-    kube = Kube(namespace=options['namespace'],
+    kube = Kube(namespace=namespace,
                 deployment_name=image,
                 printer=prompt,
                 error_printer=error_prompt)
     tag = docker_helpers.make_tag(registry, image, version)
-    yes_no = input("Deploy {} on cluster {}? [y/N] ".format(tag, kube.current_context))
-    if yes_no not in ['y', 'Y']:
-        return
     if local and not dry:
         download_requirements()
         docker_helpers.docker_image('build', tag)
@@ -197,8 +158,9 @@ def deploy(registry: str, image: str, branch: str, version: str,
 
 
 @cli.command()
-@click.option('--registry', default='???')
-@click.option('--image', default='???')
+@click.option('--registry', help='Docker registry name.', required=True)
+@click.option('--image', help='Docker image name.',
+              required=True)
 @click.option('--version', help='Version of API to build. Will replace if it'
               ' already exists.', default=None)
 def build(registry: str, image: str, version: str):
@@ -211,8 +173,9 @@ def build(registry: str, image: str, version: str):
 
 
 @cli.command()
-@click.option('--registry', default='???')
-@click.option('--image', default='???')
+@click.option('--registry', help='Docker registry name.', required=True)
+@click.option('--image', help='Docker image name.',
+              required=True)
 @click.option('--version', help='Git commit ID or branch to build and deploy.'
               ' Will replace if it already exists.', default=None)
 def push(registry: str, image: str, version: str):
@@ -224,9 +187,12 @@ def push(registry: str, image: str, version: str):
 
 
 @cli.command()
-@click.pass_obj
-def info(kube: Kube):
-    kube = Kube(namespace='???',
+@click.option('--name', help='Deployment name.', required=True)
+@click.option('--namespace', help='Namespace in the cluster.',
+              default='default')
+def info(name: str, namespace: str):
+    kube = Kube(namespace=namespace,
+                deployment_name=name,
                 printer=prompt,
                 error_printer=error_prompt)
     kube.info()
