+++ twyla/kubedeploy/prompt.py
@@ -1,0 +1,18 @@
+import colorama
+import sys
+
+# Use a colorized prompt to differentiate output of this script from output
+# that is generated by called programs and libraries
+colorama.init(autoreset=True)
+PROMPT = '>> '
+
+
+def prompt(msg: str, indent: int=0):
+    indentation = ' ' * indent
+    sys.stdout.write(colorama.Fore.GREEN + PROMPT + indentation)
+    print(msg)
+
+
+def error_prompt(msg: str):
+    sys.stdout.write(colorama.Fore.RED + PROMPT)
+    print(msg)
